pyproject.toml
[project]
name = "openvaia"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
description = "Dockerized multi-agent AI platform"

dependencies = [
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "dbmate-bin>=2.31",
    "gradio>=5.0",
    "pandas>=2.0",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.ruff]
target-version = "py312"
line-length = 120
